% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_structure.R
\name{copy_model_structure}
\alias{copy_model_structure}
\title{Copy ModelStructure.xlsx to current directory}
\usage{
copy_model_structure(dest_path = ".")
}
\arguments{
\item{dest_path}{Destination path (default: current directory)}
}
\description{
Copy ModelStructure.xlsx to current directory
}
