% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_structure.R
\name{get_model_structure_path}
\alias{get_model_structure_path}
\title{Get path to ModelStructure.xlsx template}
\usage{
get_model_structure_path()
}
\description{
Get path to ModelStructure.xlsx template
}
