% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{MatchTable}
\alias{MatchTable}
\title{MatchTable}
\usage{
MatchTable(
  Table,
  Char1 = NULL,
  Char2 = NULL,
  Char3 = NULL,
  Char4 = NULL,
  Char5 = NULL
)
}
\arguments{
\item{Table}{Name of the table}

\item{Char1}{First character string to matrix}

\item{Char2}{Second character string to matrix}

\item{Char3}{Third character string to matrix}

\item{Char4}{Fourth character string to matrix}

\item{Char5}{Fifth character string to matrix}
}
\value{
vector of matching line indices
}
\description{
This function finds the lines in a table that matches strings
}
\examples{
\dontrun{
}

}
