% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadMaterial.R
\name{ReadStarterFile}
\alias{ReadStarterFile}
\title{Parse STARTER.DAT File}
\usage{
ReadStarterFile(StarterFile)
}
\arguments{
\item{StarterFile}{Data frame from read.table() of STARTER.DAT}
}
\value{
List containing:
\itemize{
\item DataFileName - Path to DATA.DAT
\item ControlFileName - Path to CONTROL.DAT
\item SelexFileName - Path to SELEXSPEC.DAT
\item RetainFileName - Path to RETAINSPEC.DAT
\item RecruitFileName - Path to RECRUITSPEC.DAT
\item GrowthFileName - Path to GROWTHSPEC.DAT
\item MoveFileName - Path to MOVESPEC.DAT
\item TagFileName - Path to TAGFILE.TXT
\item PropFFileName - Path to PROPORTIONS.TXT
\item ProjectionsFileName - Path to PROJECTIONS.DAT
\item MaxPhase - Maximum estimation phase
}
}
\description{
Internal function to read and parse the STARTER.DAT file which contains
file paths for all other model input files.
}
\keyword{internal}
