% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadMaterial.R
\name{testindex}
\alias{testindex}
\title{Test Index Value in Data Matrix}
\usage{
testindex(Data, Row, Col, Ind)
}
\arguments{
\item{Data}{Data frame or matrix to check}

\item{Row}{Row index to check}

\item{Col}{Column index to check}

\item{Ind}{Expected index value (will be compared to Ind-1)}
}
\value{
NULL. Stops with error if validation fails.
}
\description{
Internal function to validate index values in pointer matrices during data parsing.
}
\keyword{internal}
