% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMuLT-package.R
\docType{package}
\name{IMuLT-package}
\alias{IMuLT}
\alias{IMuLT-package}
\title{IMuLT: Integrated Model using Length Transition (IMuLT)}
\description{
A package to load, control and run the stock assessment model IMuLT.
}
\author{
\strong{Maintainer}: Simon de Lestang \email{simon.delestang@dpird.wa.gov.au}

Authors:
\itemize{
  \item Andre Punt
}

}
\keyword{internal}
